Refactored train.py and predict.py by adding components to modules.py
